Move nalgebra-macros tests to nalgebra tests
By testing matrix!, stack! macros etc. in nalgebra, we ensure that
these macros are used in the same way that users will be using them.
