--- nalgebra-macros/tests/trybuild/stack_empty_col.stderr
@@ -1,7 +1,0 @@
-error: At least one element in each column must be an expression of type `Matrix`
- --> tests/trybuild/stack_empty_col.rs:5:5
-  |
-5 |     stack![0, m];
-  |     ^^^^^^^^^^^^
-  |
-  = note: this error originates in the macro `stack` (in Nightly builds, run with -Z macro-backtrace for more info)
