+++ tests/macros/trybuild/dmatrix_mismatched_dimensions.stderr
