+++ tests/macros/trybuild/stack_empty_row.stderr
@@ -1,0 +1,7 @@
+error: At least one element in each row must be an expression of type `Matrix`
+ --> $DIR/stack_empty_row.rs:5:5
+  |
+5 |     stack![0; m];
+  |     ^^^^^^^^^^^^
+  |
+  = note: this error originates in the macro `stack` (in Nightly builds, run with -Z macro-backtrace for more info)
